Simplify database preparation
